include: Move the internal headers to src.
This means that "the public API" and "what is in the
'include' directory" are now finally the same thing.
