+++ src/core-internal.h
