+++ src/profiling-internal.h
