+++ src/space-internal.h
